pyproject.toml
[project]
name = "firerisk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Yosafe Fesaha Oqbamecail", email = "581515@stud.hvl.no" }
]
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "dynamic-frcm==1.0.2",
    "fastapi[standard]>=0.129.0",
    "jinja2>=3.1.6",
    "pandas>=3.0.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-keycloak>=7.1.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]

[project.scripts]
firerisk = "firerisk:main"
firerisk-api = "firerisk.api_keycloak:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=7.4.4",
]
